pyproject.toml
[project]
name = "security-report-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.18.0",
    "litellm>=1.80.0",
    "pydantic>=2.0.0",
]
